+++ react/Settings.tsx
@@ -1,0 +1,18 @@
+import React, { Component } from 'react'
+import { withRuntimeContext } from 'render'
+
+class Settings extends Component {
+  public componentDidMount() {
+    const { runtime: { navigate } } = this.props
+
+    navigate({ to: '/admin/apps/vtex.store/setup' })
+  }
+
+  public render() {
+    return (
+      null
+    )
+  }
+}
+
+export default withRuntimeContext(Settings)