+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.0.0] - 2018-08-17
+### Added
+- Renaming repository.
+- Rendering store inside iframe.
+
 ## [1.14.2] - 2018-08-10
 
 ## [1.14.1] - 2018-08-06
@@ -14,6 +19,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.14.0] - 2018-08-06
 
 ## [1.14.0-beta] - 2018-08-03
+
+### Fixed
+- Rename to admin-pages.
 
 ## [1.13.1] - 2018-07-25
 ### Added